Add load_creds benchmark
