+++ examples/README.md
@@ -1,0 +1,1 @@
+These are small programs to benchmark some individual components.
