+++ src/main.rs
@@ -6,56 +6,20 @@ extern crate atty;
 extern crate error_chain;
 
 use badtouch::args;
-use badtouch::ctx;
 use badtouch::fsck;
-use badtouch::keyboard;
-use badtouch::pb;
-use badtouch::scheduler;
+use badtouch::utils;
+use badtouch::pb::ProgressBar;
+use badtouch::scheduler::{Scheduler, Attempt, Msg};
+use badtouch::keyboard::{Keyboard, Key};
 
-use pb::ProgressBar;
 use error_chain::ChainedError;
 use colored::*;
-use scheduler::{Scheduler, Attempt, Msg};
-use keyboard::{Keyboard, Key};
 use std::thread;
-use std::fs::{self, File};
-use std::sync::Arc;
+use std::fs::File;
 use std::time::Instant;
-use std::io::{self, BufReader};
 use std::io::prelude::*;
 use badtouch::errors::{Result, ResultExt};
 
-
-fn load_list(path: &str) -> Result<Vec<Arc<String>>> {
-    let f = File::open(path)?;
-    let file = BufReader::new(&f);
-    let lines: io::Result<_> = file.lines()
-            .map(|x| x.map(|x| Arc::new(x)))
-            .collect();
-    Ok(lines?)
-}
-
-fn load_scripts(paths: Vec<String>) -> Result<Vec<Arc<ctx::Script>>> {
-    let mut scripts = Vec::new();
-
-    for path in paths {
-        let meta = fs::metadata(&path)?;
-
-        if meta.is_dir() {
-            for path in fs::read_dir(path)? {
-                let path = path?.path();
-                let path = path.to_str().unwrap();
-                let script = Arc::new(ctx::Script::load(path)?);
-                scripts.push(script);
-            }
-        } else {
-            let script = Arc::new(ctx::Script::load(&path)?);
-            scripts.push(script);
-        }
-    }
-
-    Ok(scripts)
-}
 
 enum Report {
     Some(File),
@@ -93,11 +57,11 @@ macro_rules! info {
 }
 
 fn setup_dictionary_attack(pool: &mut Scheduler, args: args::Dict) -> Result<usize> {
-    let users = load_list(&args.users).chain_err(|| "failed to load users")?;
+    let users = utils::load_list(&args.users).chain_err(|| "failed to load users")?;
     info!("[+]", "loaded {} users", users.len());
-    let passwords = load_list(&args.passwords).chain_err(|| "failed to load passwords")?;
+    let passwords = utils::load_list(&args.passwords).chain_err(|| "failed to load passwords")?;
     info!("[+]", "loaded {} passwords", passwords.len());
-    let scripts = load_scripts(args.scripts).chain_err(|| "failed to load scripts")?;
+    let scripts = utils::load_scripts(args.scripts).chain_err(|| "failed to load scripts")?;
     info!("[+]", "loaded {} scripts", scripts.len());
 
     let attempts = users.len() * passwords.len() * scripts.len();
@@ -116,20 +80,9 @@ fn setup_dictionary_attack(pool: &mut Scheduler, args: args::Dict) -> Result<usi
 }
 
 fn setup_credential_confirmation(pool: &mut Scheduler, args: args::Creds) -> Result<usize> {
-    let creds = load_list(&args.creds)
-                    .chain_err(|| "failed to load creds")?
-                    .into_iter()
-                    .map(|x| {
-                        if let Some(idx) = x.find(":") {
-                            let (user, password) = x.split_at(idx);
-                            Ok((Arc::new(user.to_owned()), Arc::new(password[1..].to_owned())))
-                        } else {
-                            Err(format!("invalid list format: {:?}", x).into())
-                        }
-                    })
-                    .collect::<Result<Vec<_>>>()?;
+    let creds = utils::load_creds(&args.creds)?;
     info!("[+]", "loaded {} credentials", creds.len());
-    let scripts = load_scripts(args.scripts).chain_err(|| "failed to load scripts")?;
+    let scripts = utils::load_scripts(args.scripts).chain_err(|| "failed to load scripts")?;
     info!("[+]", "loaded {} scripts", scripts.len());
 
     let attempts = creds.len() * scripts.len();
